Corrected some errors
