+++ Assets/Scripts/MVC/Tank.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e9d718cbeb1434a67853c6e79b940465
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
