+++ Assets/Scripts/MVC/Tank/TankController.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TankController
+{
+    public TankModel TankModel { get; }
+    public TankView TankView { get; }
+    public TankController(TankModel tankModel, TankView tankPrefab)
+    {
+        TankModel = tankModel;
+
+        TankView = GameObject.Instantiate<TankView>(tankPrefab);
+        Debug.Log("Tank View Created");
+    }
+}
