+++ Assets/Scripts/MVC/Tank/TankController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ad25647a8a764d72842a8cec8b4efdd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
