+++ Assets/Scripts/MVC/Tank/TankService.cs
@@ -7,7 +7,12 @@ public class TankService : GenericSingleton<TankService>
     public TankView tankView;
     private void Start()
     {
+        CreateTank();
+    }
+    private TankController CreateTank()
+    {
         TankModel model = new TankModel(2, 100f);
         TankController tank = new TankController(model, tankView);
+        return tank;
     }
 }
