+++ Assets/Scripts/MVC/Tank/TankView.cs
