+++ Assets/Scripts/Singleton/TankControllerAndroid.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 834624795e9c44fd0b6477821d5c116b
+guid: 6c2c6b14f190046bfbf62a9951b02bbe
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
